# user/user-back/docs/swagger.yaml
definitions:
  controller.RegisterDTO:
    properties:
      first_name:
        type: string
      last_name:
        type: string
      password:
        type: string
      patronymic:
        type: string
      phone:
        type: string
    type: object
  controller.TokensRes:
    properties:
      access_token:
        type: string
      refresh_token:
        type: string
    type: object
  models.User:
    properties:
      created_at:
        type: string
      first_name:
        type: string
      id:
        type: integer
      last_name:
        type: string
      patronymic:
        type: string
      phone:
        type: string
      role:
        type: string
      updated_at:
        type: string
    type: object
info:
  contact: {}
paths:
  /login:
    get:
      description: Возвращает токены доступа при успешной авторизации
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/controller.TokensRes'
        "500":
          description: Internal Server Error
          schema:
            type: string
      summary: Авторизация пользователя
      tags:
      - auth
  /register:
    post:
      consumes:
      - application/json
      description: Создает нового пользователя и возвращает токены доступа
      parameters:
      - description: Данные нового пользователя
        in: body
        name: user
        required: true
        schema:
          $ref: '#/definitions/controller.RegisterDTO'
      produces:
      - application/json
      responses:
        "201":
          description: Created
          schema:
            $ref: '#/definitions/controller.TokensRes'
        "400":
          description: Invalid JSON
          schema:
            type: string
        "500":
          description: Internal Server Error
          schema:
            type: string
      summary: Регистрация пользователя
      tags:
      - auth
  /users:
    get:
      description: Возвращает массив всех зарегистрированных пользователей
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            items:
              $ref: '#/definitions/models.User'
            type: array
        "500":
          description: Internal Server Error
          schema:
            type: string
      summary: Получить список пользователей
      tags:
      - users
    post:
      consumes:
      - application/json
      description: Регистрирует нового пользователя. Поля name, email и password обязательны.
      parameters:
      - description: Данные нового пользователя
        in: body
        name: user
        required: true
        schema:
          $ref: '#/definitions/models.User'
      produces:
      - application/json
      responses:
        "201":
          description: Created
          schema:
            $ref: '#/definitions/models.User'
        "400":
          description: Invalid JSON или пустые обязательные поля
          schema:
            type: string
        "500":
          description: Internal Server Error
          schema:
            type: string
      summary: Создать пользователя
      tags:
      - users
  /users/{id}:
    delete:
      description: Удаляет запись о пользователе из базы данных по его ID
      parameters:
      - description: Идентификатор пользователя
        in: path
        name: id
        required: true
        type: integer
      produces:
      - application/json
      responses:
        "200":
          description: '{"status": "deleted"}'
          schema:
            additionalProperties:
              type: string
            type: object
        "400":
          description: Invalid ID
          schema:
            type: string
        "500":
          description: Internal Server Error
          schema:
            type: string
      summary: Удалить пользователя
      tags:
      - users
    get:
      description: Возвращает профиль пользователя по его числовому идентификатору
      parameters:
      - description: Идентификатор пользователя
        in: path
        name: id
        required: true
        type: integer
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/models.User'
        "400":
          description: Invalid ID
          schema:
            type: string
        "404":
          description: User Not Found
          schema:
            type: string
      summary: Получить пользователя по ID
      tags:
      - users
    put:
      consumes:
      - application/json
      description: Полностью обновляет информацию о пользователе по его ID
      parameters:
      - description: Идентификатор пользователя
        in: path
        name: id
        required: true
        type: integer
      - description: Новые данные пользователя
        in: body
        name: user
        required: true
        schema:
          $ref: '#/definitions/models.User'
      produces:
      - application/json
      responses:
        "200":
          description: '{"status": "updated"}'
          schema:
            additionalProperties:
              type: string
            type: object
        "400":
          description: Invalid ID или Invalid JSON
          schema:
            type: string
        "500":
          description: Internal Server Error
          schema:
            type: string
      summary: Обновить профиль пользователя
      tags:
      - users
  /users/{id}/role:
    patch:
      consumes:
      - application/json
      description: 'Частичное обновление: изменяет только роль пользователя (например,
        admin, user)'
      parameters:
      - description: Идентификатор пользователя
        in: path
        name: id
        required: true
        type: integer
      - description: JSON вида {
        in: body
        name: role
        required: true
        schema:
          type: string
      produces:
      - application/json
      responses:
        "200":
          description: '{"status": "role updated"}'
          schema:
            additionalProperties:
              type: string
            type: object
        "400":
          description: Invalid ID, Invalid JSON или пустая роль
          schema:
            type: string
        "500":
          description: Internal Server Error
          schema:
            type: string
      summary: Обновить роль пользователя
      tags:
      - users
  /users/email/{email}:
    get:
      description: Ищет и возвращает профиль пользователя по его электронной почте
      parameters:
      - description: Email адрес пользователя
        in: path
        name: email
        required: true
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/models.User'
        "400":
          description: Email is required
          schema:
            type: string
        "404":
          description: User Not Found
          schema:
            type: string
      summary: Найти пользователя по Email
      tags:
      - users
swagger: "2.0"